Move data prep functions to train module
