+++ src/planetsca/__init__.py
@@ -1,8 +1,7 @@
-from . import data_preparation, download, predict, search, train
+from . import download, predict, search, train
 
 __all__ = [
     "download",
-    "data_preparation",
     "train",
     "predict",
     "search",
